+++ mayan/apps/document_indexing/apps.py
@@ -34,7 +34,8 @@ from .links import (
 from .licenses import *  # NOQA
 from .permissions import (
     permission_document_indexing_create, permission_document_indexing_delete,
-    permission_document_indexing_edit, permission_document_indexing_view
+    permission_document_indexing_edit, permission_document_indexing_rebuild,
+    permission_document_indexing_view
 )
 from .queues import *  # NOQA
 from .widgets import get_instance_link, index_instance_item_link, node_level
@@ -72,6 +73,7 @@ class DocumentIndexingApp(MayanAppConfig):
                 permission_document_indexing_create,
                 permission_document_indexing_delete,
                 permission_document_indexing_edit,
+                permission_document_indexing_rebuild,
                 permission_document_indexing_view,
             )
         )